Refactor moving test files.
