+++ src/test/java/jp/aha/oretama/typoFixer/service/TypoCheckerServiceTest.java
@@ -1,4 +1,4 @@
-package jp.aha.oretama.typoFixer;
+package jp.aha.oretama.typoFixer.service;
 
 import jp.aha.oretama.typoFixer.model.Diff;
 import jp.aha.oretama.typoFixer.service.TypoCheckerService;
