+++ src/test/java/jp/aha/oretama/typoFixer/service/TypoModifierServiceTest.java
@@ -1,4 +1,4 @@
-package jp.aha.oretama.typoFixer;
+package jp.aha.oretama.typoFixer.service;
 
 import jp.aha.oretama.typoFixer.model.Event;
 import jp.aha.oretama.typoFixer.model.Modification;
